requirements.txt
cot-reports
dash
dash-bootstrap-components
duckdb
fredapi
loguru
matplotlib
numpy
pandas
pandas-datareader
plotly
psycopg2-binary
python-dotenv
requests
rich
schedule
scikit-learn
scipy
statsmodels
yfinance
